+++ src/design/plone/contenttypes/browser/utils/check_servizi.py
@@ -1,14 +1,16 @@
 # -*- coding: utf-8 -*-
+
 from DateTime import DateTime
 from design.plone.contenttypes.utils import text_in_block
 from openpyxl import Workbook
+from openpyxl.cell.cell import ILLEGAL_CHARACTERS_RE
 from openpyxl.styles import Alignment
 from openpyxl.styles import Font
 from openpyxl.styles import PatternFill
 from openpyxl.utils import get_column_letter
+
 from plone import api
 from Products.Five import BrowserView
-
 import io
 
 
@@ -288,7 +290,7 @@ class DownloadCheckServizi(CheckServizi):
             if have_url:
                 url = row.pop()
 
-            sheet.append(row)
+            sheet.append([ILLEGAL_CHARACTERS_RE.sub(r" ", x) for x in row])
 
             if row[0] == "Titolo":
                 for index, cell in enumerate(sheet[i]):